backend/requirements.txt
aioredis==1.3.1
amqp==5.0.9
asgiref==3.5.0
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
boto3==1.20.7
botocore==1.23.7
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.12
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coolsms-python-sdk==2.0.3
cryptography==36.0.1
daphne==3.0.2
Django==3.2.9
django-cors-headers==3.10.0
django-crontab==0.7.1
django-filter==21.1
django-hosts==4.0
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.21.0
feedparser==6.0.8
future==0.18.2
gevent==21.8.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-resources==5.4.0
incremental==21.3.0
inflection==0.5.1
jmespath==0.10.0
jsonschema==4.4.0
kombu==5.2.3
lxml==4.8.0
msgpack==1.0.3
numpy==1.22.3
oauthlib==3.2.0
outcome==1.1.0
pandas==1.4.2
Pillow==8.4.0
prompt-toolkit==3.0.28
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-twitter==3.5
pytrends==4.8.0
pytz==2021.3
PyYAML==6.0
redis==3.5.3
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.5.1
selenium==4.1.5
service-identity==21.1.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
trio==0.20.0
trio-websocket==0.9.2
tweepy==4.9.0
Twisted==22.1.0
txaio==21.2.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
webdriver-manager==3.5.4
wsproto==1.1.0
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
